refact/front/GoHomeBtn: moves to src/components/buttons/GoHomeBtn.jsx
